+++ string/ft_strchr.c
@@ -20,6 +20,6 @@ char	*ft_strchr(const char *s, int c)
 	while (s[i] && s[i] != c)
 		i++;
 	if (s[i] == c)
-		return ((char *)&s[i]);
+		return ((char *)(size_t)&s[i]);
 	return (NULL);
 }
