+++ string/ft_strnstr.c
@@ -17,12 +17,12 @@ char	*ft_strnstr(const char *s1, const char *s2, size_t n)
 	const size_t len = ft_strlen(s2);
 
 	if (len == 0)
-		return ((char *)s1);
+		return ((char *)(size_t)s1);
 	while (*s1 && n >= len)
 	{
 		n--;
 		if (ft_strncmp(s1, s2, len) == 0)
-			return ((char *)s1);
+			return ((char *)(size_t)s1);
 		s1++;
 	}
 	return (NULL);
